Use floats for positioning

+++ app/src/main/java/com/danosoftware/galaxyforce/flightpath/paths/PathPoint.java
@@ -1,30 +1,29 @@
 package com.danosoftware.galaxyforce.flightpath.paths;
 
 /**
- * Represents a point object (an x-y co-ordinate).
- * Normally used to represent a position on a path.
+ * Represents a point object (an x-y co-ordinate). Normally used to represent a position on a path.
  */
 public class PathPoint {
 
-    private final int x;
-    private final int y;
-    private final int angle;
+  private final float x;
+  private final float y;
+  private final float angle;
 
-    public PathPoint(int x, int y, int angle) {
-        this.x = x;
-        this.y = y;
-        this.angle = angle;
-    }
+  public PathPoint(float x, float y, float angle) {
+    this.x = x;
+    this.y = y;
+    this.angle = angle;
+  }
 
-    public int getX() {
-        return x;
-    }
+  public float getX() {
+    return x;
+  }
 
-    public int getY() {
-        return y;
-    }
+  public float getY() {
+    return y;
+  }
 
-    public int getAngle() {
-        return angle;
-    }
+  public float getAngle() {
+    return angle;
+  }
 }
