+++ app/src/main/java/com/danosoftware/galaxyforce/sprites/game/aliens/implementations/FollowerAlien.java
@@ -10,7 +10,6 @@ import com.danosoftware.galaxyforce.sprites.game.behaviours.powerup.PowerUpBehav
 import com.danosoftware.galaxyforce.sprites.game.behaviours.spawn.SpawnBehaviourFactory;
 import com.danosoftware.galaxyforce.sprites.game.behaviours.spinner.SpinningBehaviourFactory;
 import com.danosoftware.galaxyforce.waves.config.aliens.types.FollowerConfig;
-
 import lombok.Builder;
 import lombok.NonNull;
 
@@ -26,36 +25,36 @@ public class FollowerAlien extends AbstractAlienFollower implements IAlienFollow
 
     @Builder
     public FollowerAlien(
-            @NonNull final ExplosionBehaviourFactory explosionFactory,
-            @NonNull final SpawnBehaviourFactory spawnFactory,
-            @NonNull final SpinningBehaviourFactory spinningFactory,
-            @NonNull final PowerUpBehaviourFactory powerUpFactory,
-            @NonNull final FireBehaviourFactory fireFactory,
-            @NonNull final HitBehaviourFactory hitFactory,
-            @NonNull final FollowerConfig alienConfig,
-            final PowerUpType powerUpType,
-            @NonNull final Integer xStart,
-            @NonNull final Integer yStart) {
+        @NonNull final ExplosionBehaviourFactory explosionFactory,
+        @NonNull final SpawnBehaviourFactory spawnFactory,
+        @NonNull final SpinningBehaviourFactory spinningFactory,
+        @NonNull final PowerUpBehaviourFactory powerUpFactory,
+        @NonNull final FireBehaviourFactory fireFactory,
+        @NonNull final HitBehaviourFactory hitFactory,
+        @NonNull final FollowerConfig alienConfig,
+        final PowerUpType powerUpType,
+        @NonNull final Float xStart,
+        @NonNull final Float yStart) {
 
-        super(
-                alienConfig.getAlienCharacter(),
-                alienConfig.getAlienCharacter().getAnimation(),
-                xStart,
-                yStart,
-                alienConfig.getEnergy(),
-                fireFactory.createFireBehaviour(
-                        alienConfig.getMissileConfig()),
-                powerUpFactory.createPowerUpBehaviour(
-                        powerUpType),
-                spawnFactory.createSpawnBehaviour(
-                        alienConfig.getSpawnConfig()),
-                hitFactory.createHitBehaviour(
-                        alienConfig.getAlienCharacter().getHitAnimation()),
-                explosionFactory.createExplosionBehaviour(
-                        alienConfig.getExplosionConfig(),
-                        alienConfig.getAlienCharacter()),
-                spinningFactory.createSpinningBehaviour(
-                        alienConfig.getSpinningConfig(),
+      super(
+          alienConfig.getAlienCharacter(),
+          alienConfig.getAlienCharacter().getAnimation(),
+          xStart,
+          yStart,
+          alienConfig.getEnergy(),
+          fireFactory.createFireBehaviour(
+              alienConfig.getMissileConfig()),
+          powerUpFactory.createPowerUpBehaviour(
+              powerUpType),
+          spawnFactory.createSpawnBehaviour(
+              alienConfig.getSpawnConfig()),
+          hitFactory.createHitBehaviour(
+              alienConfig.getAlienCharacter().getHitAnimation()),
+          explosionFactory.createExplosionBehaviour(
+              alienConfig.getExplosionConfig(),
+              alienConfig.getAlienCharacter()),
+          spinningFactory.createSpinningBehaviour(
+              alienConfig.getSpinningConfig(),
                         alienConfig.getSpeed()),
                 alienConfig
                         .getSpeed()
