+++ app/src/main/java/com/danosoftware/galaxyforce/sprites/game/missiles/aliens/AlienMissileDownwards.java
@@ -11,22 +11,22 @@ public class AlienMissileDownwards extends AbstractAlienMissile {
     private final int missileSpeed;
 
     public AlienMissileDownwards(
-            int xStart,
-            int yStart,
-            final Animation animation,
-            final AlienMissileSpeed missileSpeed) {
-        super(
-                animation,
-                xStart,
-                yStart);
-        this.missileSpeed = -missileSpeed.getSpeed();
+        float xStart,
+        float yStart,
+        final Animation animation,
+        final AlienMissileSpeed missileSpeed) {
+      super(
+          animation,
+          xStart,
+          yStart);
+      this.missileSpeed = -missileSpeed.getSpeed();
     }
 
     @Override
     public void animate(float deltaTime) {
         super.animate(deltaTime);
 
-        moveYByDelta((int) (missileSpeed * deltaTime));
+      moveYByDelta(missileSpeed * deltaTime);
 
         // if missile is now off screen then destroy it
         if (offScreenBottom(this)) {
