+++ app/src/main/java/com/danosoftware/galaxyforce/sprites/game/missiles/aliens/AlienMissileRotated.java
@@ -21,68 +21,68 @@ import com.danosoftware.galaxyforce.view.Animation;
  */
 public class AlienMissileRotated extends AbstractAlienMissile {
 
-    private static final float DEGREES_PER_PI_RADIANS = (float) (180f / Math.PI);
+  private static final float DEGREES_PER_PI_RADIANS = (float) (180f / Math.PI);
 
-    // offset applied to x and y every move
-    private final int xDelta;
-    private final int yDelta;
+  // offset applied to x and y every move
+  private final float xDelta;
+  private final float yDelta;
 
-    // created rotated missile aimed at base
-    public AlienMissileRotated(
-            int xStart,
-            int yStart,
-            final Animation animation,
-            final AlienMissileSpeed missileSpeed,
-            final IBasePrimary base) {
+  // created rotated missile aimed at base
+  public AlienMissileRotated(
+      float xStart,
+      float yStart,
+      final Animation animation,
+      final AlienMissileSpeed missileSpeed,
+      final IBasePrimary base) {
 
-        super(
-                animation,
-                xStart,
-                yStart);
+    super(
+        animation,
+        xStart,
+        yStart);
 
-        // calculate angle from missile position to base
-        final AlienMissileRotateCalculation calculation = calculateAngle(this, base);
-        final float angle = calculation.getAngle();
-        rotate(calculation.getRotation());
+    // calculate angle from missile position to base
+    final AlienMissileRotateCalculation calculation = calculateAngle(this, base);
+    final float angle = calculation.getAngle();
+    rotate(calculation.getRotation());
 
-        // calculate the deltas to be applied each move
-        this.xDelta = (int) (missileSpeed.getSpeed() * (float) Math.cos(angle));
-        this.yDelta = (int) (missileSpeed.getSpeed() * (float) Math.sin(angle));
-    }
+    // calculate the deltas to be applied each move
+    this.xDelta = missileSpeed.getSpeed() * (float) Math.cos(angle);
+    this.yDelta = missileSpeed.getSpeed() * (float) Math.sin(angle);
+  }
 
-    // created rotated missile of supplied angle
-    public AlienMissileRotated(
-            int xStart,
-            int yStart,
-            final Animation animation,
-            final AlienMissileSpeed missileSpeed,
-            final float angle) {
+  // created rotated missile of supplied angle
+  public AlienMissileRotated(
+      float xStart,
+      float yStart,
+      final Animation animation,
+      final AlienMissileSpeed missileSpeed,
+      final float angle) {
 
-        super(
-                animation,
-                xStart,
-                yStart);
+    super(
+        animation,
+        xStart,
+        yStart);
 
-        // calculate sprite rotation for wanted angle
-        rotate((calculateRotation(angle)));
+    // calculate sprite rotation for wanted angle
+    rotate((calculateRotation(angle)));
 
-        // calculate the deltas to be applied each move
-        this.xDelta = (int) (missileSpeed.getSpeed() * (float) Math.cos(angle));
-        this.yDelta = (int) (missileSpeed.getSpeed() * (float) Math.sin(angle));
-    }
+    // calculate the deltas to be applied each move
+    this.xDelta = missileSpeed.getSpeed() * (float) Math.cos(angle);
+    this.yDelta = missileSpeed.getSpeed() * (float) Math.sin(angle);
+  }
 
     @Override
     public void animate(float deltaTime) {
-        super.animate(deltaTime);
+      super.animate(deltaTime);
 
-        // move missile by calculated deltas
-        moveByDelta(
-                (int) (xDelta * deltaTime),
-                (int) (yDelta * deltaTime));
+      // move missile by calculated deltas
+      moveByDelta(
+          xDelta * deltaTime,
+          yDelta * deltaTime);
 
-        // if missile is now off screen then destroy it
-        if (offScreenAnySide(this)) {
-            destroy();
-        }
+      // if missile is now off screen then destroy it
+      if (offScreenAnySide(this)) {
+        destroy();
+      }
     }
 }
